+++ react/components/SKUSelector/index.tsx
@@ -299,6 +299,11 @@ const SKUSelectorContainer: FC<Props> = ({
       isMainAndImpossible,
       possibleItems,
     }) => {
+      dispatch({
+        type: 'SET_LOADING_ITEM',
+        args: { loadingItem: true },
+      })
+
       // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
       const isRemoving = selectedVariations![variationName] === variationValue
       const newSelectedVariation = !isMainAndImpossible