+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- `onClickItem` from SKUSelector to dispatch `SET_LOADING_ITEM` action. 
 
 ## [3.149.0] - 2021-07-15
 ### Added